Add files via upload
Write a Java program to compare two numbers. Go to the editor

Input Data:

Input first integer: 25

Input second integer: 39


25 != 39
25 < 39
25 <= 39


Click me to see the solution

Write a Java program and compute the sum of the digits of an integer. Go to the editor

Input Data:

Input an integer: 25


The sum of the digits is: 7


Click me to see the solution

Write a Java program to compute the area of a hexagon. Go to the editor

Area of a hexagon = (6 * s^2)/(4*tan(π/6))

where s is the length of a side

Input Data:

Input the length of a side of the hexagon: 6


The area of the hexagon is: 93.53074360871938


Click me to see the solution

Write a Java program to compute the area of a polygon. Go to the editor

Area of a polygon = (n*s^2)/(4*tan(π/n))

where n is n-sided polygon and s is the length of a side

Input Data:

Input the number of sides on the polygon: 7

Input the length of one of the sides: 6


The area is: 130.82084798405722


Click me to see the solution

Write a Java program to compute the distance between two points on the surface of earth. Go to the editor

Distance between the two points [ (x1,y1) & (x2,y2)]

d = radius * arccos(sin(x1) * sin(x2) + cos(x1) * cos(x2) * cos(y1 - y2))

Radius of the earth r = 6371.01 Kilometers

Input Data:

Input the latitude of coordinate 1: 25

Input the longitude of coordinate 1: 35

Input the latitude of coordinate 2: 35.5

Input the longitude of coordinate 2: 25.5


The distance between those points is: 1480.0848451069087 km


Click me to see the solution

Write a Java program to reverse a string. Go to the editor

Input Data:

Input a string: The quick brown fox


Reverse string: xof nworb kciuq ehT


Click me to see the solution

Write a Java program to count the letters, spaces, numbers and other characters of an input string. Go to the editor


The string is : Aa kiu, I swd skieo 236587. GH kiu: sieo?? 25.33
letter: 23
space: 9
number: 10
other: 6


Click me to see the solution

Write a Java program to create and display
unique three-digit number using 1, 2, 3, 4. Also count how many
three-digit numbers are there. Go to the editor


123
124
...

431
432
Total number of the three-digit-number is 24


Click me to see the solution

Write a Java program to list the available character sets in charset objects. Go to the editor


List of available character sets:
Big5
Big5-HKSCS
CESU-8
EUC-JP
EUC-KR
GB18030
GB2312
GBK

...

x-SJIS_0213
x-UTF-16LE-BOM
X-UTF-32BE-BOM
X-UTF-32LE-BOM
x-windows-50220
x-windows-50221
x-windows-874
x-windows-949
x-windows-950
x-windows-iso2022jp


Click me to see the solution

Write a Java program to print the ascii value of a given character. Go to the editor


The ASCII value of Z is :90


Click me to see the solution

Write a Java program to input and display your password. Go to the editor


Input your Password:
Your password was: abc@123


Click me to see the solution

Write a Java program to print the following string in a specific format (see the output). Go to the editor


Twinkle, twinkle, little star,
   How I wonder what you are!
       Up above the world so high,        
       Like a diamond in the sky.
Twinkle, twinkle, little star,
   How I wonder what you are


Click me to see the solution

Write a Java program that accepts an integer (n) and computes the value of n+nn+nnn. Go to the editor


Input number: 5
5 + 55 + 555


Click me to see the solution

Write a Java program to find the size of a specified file. Go to the editor


/home/students/abc.txt : 0 bytes
/home/students/test.txt : 0 bytes


Click me to see the solution

Write a Java program to display the system time. Go to the editor


Current Date time: Fri Jun 16 14:17:40 IST 2017


Click me to see the solution

Write a Java program to display the current date time in specific format. Go to the editor


Now: 2017/06/16 08:52:03.066


Click me to see the solution

Write a Java program to print the odd numbers from 1 to 99. Prints one number per line. Go to the editor


1
3
5
7
9
11
....

91
93
95
97
99


Click me to see the solution

Write a Java program to accept a number and
check the number is even or not. Prints 1 if the number is even or 0 if
the number is odd. Go to the editor


Input a number: 20
1


Click me to see the solution

Write a Java program to print numbers between 1 to 100 which are divisible by 3, 5 and by both. Go to the editor


Divided by 3:
3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36, 39, 42, 45, 48, 51, 54, 57
, 60, 63, 66, 69, 72, 75, 78, 81, 84, 87, 90, 93, 96, 99,

Divided by 5:
5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80, 85, 90,
95,

Divided by 3 & 5:
15, 30, 45, 60, 75, 90,