Add files via upload
Write a Java program to create the concatenation
of the two strings except removing the first character of each string.
The length of the strings must be 1 and above. Go to the editor

Test Data: Str1 = Python

Str2 = Tutorial


ythonutorial

Click me to see the solution

Write a Java program to create a new string
taking first three characters from a given string. If the length of the
given string is less than 3 use "#" as substitute characters. Go to the editor

Test Data: Str1 = " "


###

Click me to see the solution

Write a Java program to create a new string
taking first and last characters from two given strings. If the length
of either string is 0 use "#" for missing character. Go to the editor

Test Data: str1 = "Python"

str2 = " "


P#

Click me to see the solution

Write a Java program to test if 10 appears as
either the first or last element of an array of integers. The length of
the array must be greater than or equal to 2. Go to the editor



Test Data: array = 10, -20, 0, 30, 40, 60, 10
true

Click me to see the solution

Write a Java program to test if the first and
the last element of an array of integers are same. The length of the
array must be greater than or equal to 2. Go to the editor

Test Data: array = 50, -20, 0, 30, 40, 60, 10


false

Click me to see the solution

Write a Java program to test if the first or the
last element of two array of integers are same. The length of the array
must be greater than or equal to 2. Go to the editor
Test Data: array1 = 50, -20, 0, 30, 40, 60, 12

array2 = 45, 20, 10, 20, 30, 50, 11


false

Click me to see the solution

Write a Java program to create a new array of
length 2 from two arrays of integers with three elements and the new
array will contain the first and last elements from the two arrays. Go to the editor

Test Data: array1 = 50, -20, 0

array2 = 5, -50, 10


Array1: [50, -20, 0]
Array2: [5, -50, 10]
New Array: [50, 10]

Click me to see the solution

Write a Java program to test that a given array of integers of length 2 contains a 4 or a 7. Go to the editor


Original Array: [5, 7]
true

Click me to see the solution

Write a Java program to rotate an array (length 3) of integers in left direction. Go to the editor


Original Array: [20, 30, 40]
Rotated Array: [30, 40, 20]

Click me to see the solution

Write a Java program to get the larger value between first and last element of an array (length 3) of integers . Go to the editor


Original Array: [20, 30, 40]
Larger value between first and last element: 40

Click me to see the solution

Write a Java program to swap the first and last elements of an array (length must be at least 1) and create a new array. Go to the editor


Original Array: [20, 30, 40]
New array after swaping the first and last elements: [40, 30, 20]

Click me to see the solution

Write a Java program to find the largest element
between first, last, and middle values from an array of integers (even
length). Go to the editor


Original Array: [20, 30, 40, 50, 67]
Largest element between first, last, and middle values: 67

Click me to see the solution

Write a Java program to multiply corresponding elements of two arrays of integers. Go to the editor


Array1: [1, 3, -5, 4]

Array2: [1, 4, -5, -2]

Result: 1 12 25 -8

Click me to see the solution

Write a Java program to take the last three
characters from a given string and add the three characters at both the
front and back of the string. String length must be greater than three
and more. Go to the editor

Test data: "Python" will be "honPythonhon"


honPythonhon

Click me to see the solution

Write a Java program to check if a string starts with a specified word. Go to the editor
Sample Data: string1 = "Hello how are you?"


true


Click me to see the solution

Write a Java program start with an integer n,
divide n by 2 if n is even or multiply by 3 and add 1 if n is odd,
repeat the process until n = 1. Go to the editor


Click me to see the solution

Write a Java program than read an integer and
calculate the sum of its digits and write the number of each digit of
the sum in English. Go to the editor


Click me to see the solution

Write a Java program to get the current system environment and system properties. Go to the editor

Click me to see the solution

Write a Java program to check whether a security manager has already been established for the current application or not. Go to the editor

Click me to see the solution

Write a Java program to get the value of the environment variable PATH, TEMP, USERNAME. Go to the editor

Click me to see the solution

Write a Java program to measure how long some code takes to execute in nanoseconds. Go to the editor

Click me to see the solution

Write a Java program to count the number of even and odd elements in a given array of integers. Go to the editor

Click me to see the solution

Write a Java program to test if an array of
integers contains an element 10 next to 10 or an element 20 next to 20,
but not both. Go to the editor

Click me to see the solution

Write a Java program to rearrange all the
elements of a given array of integers so that all the odd numbers come
before all the even numbers. Go to the editor

Click me to see the solution

Write a Java program to create an array (length #
0) of string values. The elements will contain "0", "1", "2" … through
... n-1. Go to the editor

Click me to see the solution

Write a Java program to check if there is a 10 in a given array of integers with a 20 somewhere later in the array. Go to the editor

Click me to see the solution

Write a Java program to check if an array of
integers contains a specified number next to each other or there are two
same specified numbers separated by one element. Go to the editor

Click me to see the solution

Write a Java program to check if the value 20
appears three times and no 20's are next to each other in a given array
of integers. Go to the editor

Click me to see the solution

Write a Java program to check if a specified number appears in every pair of adjacent element of a given array of integers. Go to the editor

Click me to see the solution

Write a Java program to count the two elements of two given arrays of integers with same length, differ by 1 or less. Go to the editor

Click me to see the solution